Default wordwrite with single arg to use offset 0
The KATCP wordwrite request requires a word_offset and one or more data
values.  When writing to a single valued register, the offset is
(almost?) always 0.  It is much more intuitive to be able to use...

roach.wordwrite('reg_name', value)

...instead of...

roach.wordwrite('reg_name', 0, value)

Prior to this change, the more intuitive way did not error out, but did
not have the intuitively expected effect.