Simplify Response#payload method
The Response#payload method used to accept arguments that would be used
to transform the payload String to other (e.g. numeric) types.  This was
of little value since the user can simply call the transformation method
directly on the returned String object.

To alleviate the need to check for nil before calling the conversion
method, #payload now returns an empty String if #ok? is false or the
Response has no payload.