Change args and return type of #read and #write
RoachClient#read and #write used to take byte based offsets and counts,
but now they take word based offsets and counts.  This was changed to
match ROACH device access requirement of word alignment for offsets and
counts thereby eliminating the need for *4 or /4 conversions.  These
methods also deal with client-side native values, automatically
converting endianess as needed.

The #read method now returns an Integer if no word count was given or an
NArray.int if a word count was given.  The #write method now returns
self.

The data passed to #write can be a String (passed as-is), an
NArray in native endianess, an Array or individual object compatible
with Array#pack's "N*" format.

Both #read and #write will raise an exception if the operation fails.