Dynamically define/undefine device access methods
Dynamically define (and undefine) device access methods as the FPGA is
programmed (and de-programmed or re-programmed).

Add a #devices attribute reader that returns an Array of Strings
containing device names for the currently programmed gateware design.
This Array is empty if the design has no devices (very rare, if even
possible) or if the device is not programmed.

Add #has_device?(name) method to check whether the currently programmed
gateware has a given device name.

Add #has_key?(name) method to further flesh out some Hash-like
functionality.

Add #programmed? method to make it easy to check whether the device is
programmed.