Fix bug in read's handling of word_offset
If read is passed a word offset, but not a word count, it used to pass
the word offset as a byte offset to the underlying wordread request.
Now it always uses the underlying KATCP read request and properly
handles word based to byte based conversions regardless of how it is
called.