Make KATCP::Client#method_missing more cranky
If method_missing's request results in a Response object whose status is not
OK, raise an exception.  Previously it would silently return the response
object, which place the onus on the caller to check the status and didn't work
at all for missing assignment methods since the rvalue is returned rather than
the Response object.