Use select to enable timeout on socket reads
The socket reader thread used to use gets to read lines from the socket, but
this would hang indefinitely if the KATCP server (or its host) went
down for whatever reason.  That behavior is not very robust.

Now the socket reader thread uses the IO#select call with a timeout
value of 0.25 seconds to prevent indefinite blocking.  If two timeouts
(or an error on the socket) occur while waiting for a reply, the
connection is assumed to be dead and the reader thread puts one last
"double-bang" message on @rxq (either "!!socket-timeout" or
"!!socket-error").  Upon popping the doble-bang methods from @rxq, the
thread in the request method will close the socket and raise either
TimeoutError or SocketError as approriate.  Future requests on the same
Client object will fail with an IOError (closed stream).