Improve KATCP::Client robustness
KATCP::Client now handled timeouts, socket errors, and socket EOF by
automactically trying to reconnect once.  If successful, things proceed
as if nothing happened.  If unsuccessful, an approprite exception is
rasied, but the Client object itself is still valid.  Future requests
can be made and they will (attempt to) reconnect if necessary.

Failure to connect in the constructor does not raise an exception so
KATCP::Client objects can be created for servers that are not (yet) up.