Support newer tcpborphserver versions
Newer tcpborphserver versions no longer support bulkread and have
renamed the "status" command to be "fpgastatus".  This change determines
which commands are supported by the server at connection time and
handles things appropriately.  This adds an explicit "bulkread" method
which uses the KATCP "?read" command on newer tcpborphservers that do
not support "?bulkread".