Fix handling of typemap passed to RoachClient.new
Passing a typemap to RoachClient's constructor via the trailing options
Hash was broken.  The typemap hash was being stored in @device_typemap
AFTER the superclass constructor was called.  The superclass ctor calls
Client#connect which will dynamically create accessor methods for the
gateware devices if the FPGA is already programmed.  The problem is that
the device typemap has not yet been stored in @device_typemap, so
the dynamically created methods are made without the typemap info.  The
fix is to do away with the @device_typemap instance variable altogether
and simply leave the passed in typemap in the @opts[:typemap] value,
which is set well before anything will try to use it.