Fix TenGE use of overloaded [] and []=
TenGE overloaded the [] and []= methods to provide access to ARP table.  This
means that other methods in TenGE cannot use them to access arbitrary
locations.  To fix this, the Bram class (the superclass of TenGE) provides #get
and #set aliases for its [] and []= methods.  These aliases are used by the
methods in TenGE that cannot use [] and []=.