Make device_type map support user specified class
Instead of supporting a finite set of predefined classes in
device_typemap, we now support any class the user specifies.  The device
type_map entry could contain:

  :histo_a1_0 => [Histo, histo_a1],
  :histo_a1_4 => :skip

For a device named "histo_a1_0", this would create a method of the same
name that returns a Histo object.  The Histo constructor must accept a
KATCP::client instance and the name of the referenced device.  In the
example above, the alias "histo_a1" would be created for the
"histo_a1_0" method.  Furthermore, no method would be created for the
histo_a1_4 object since that key has the value :skip in the typemap.