Change connection timeout mechanism
This used to use non-blocking connects, but that did not work well on
Windows (and maybe JRuby).  The new approach uses the Timeout module
from the Ruby Standard Library.  This new technique was borrowed from
the MongoDB project.  See this commit for more info:

https://github.com/mongodb/mongo-ruby-driver/commit/65a1ecc42d4ab4c633e6487b8cb932637e7505c1

The socket timeout can now be given via the options Hash of the
KATCP::Client constructor using the :socket_timeout key.  The value
should be a (possibly floating point) number.  If not given, the default
is DEFAULT_SOCKET_TIMEOUT (currently 0.25 seconds).

The remote host and port are packed into a sockaddr String early in the
constructor now so that getaddrinfo errors (e.g. unkonwn host errors)
raise exceptions as early as possible.