Fix usage of select call
The select call is used to wait for the TCP socket to become readable.
The original usage of the select call also waited for an "exceptional
condition" on the TCP socket.  This was erroneously interpreted as
meaning that some sort of asynchronous error had occurred on the socket
(e.g. connection dropped).  This is not the correct interpretation of
the "exceptions" condition.  TCP sockets with "exceptions" simply have
urgent/oob data that can be read.  KATCP does not use this so in theory
it should not really cause problems.  On Linux and Mac OS X that is
true, but on Windows the select call returns the socket as both readable
and with an "exception".  Since we don't care about these "execptional"
conditions (a problem with the socket will cause read to fail), we now
only wait for the socket to become readable.