Add DEVICE_TYPEMAP to RoachClient class
Also change how typemap passed to constructor is used.  It is now merged
into the Hash returned by device_typemap.  This can have side effects
that might be unwanted if device_typemap returns a Hash that is
referenced by a class constant as in:

  class Foo < KATCP::RoachClient
    DEVICE_TYPEMAP = {...}
    def device_typemap
      DEVICE_TYPEMAP
    end
  end

If the above case, calling the constructor with an optional typemap will
modify the Hash reference by the class constant.  To avoid that, the
recommended practice is to define the above class like this:

  class Foo < KATCP::RoachClient
    DEVICE_TYPEMAP = {...}
    def device_typemap
      @device_typemap ||= DEVICE_TYPEMAP.dup
    end
  end