Fix handling of options Hash in constructor
The constructors of KATCP::Roach and KATCP::RoachClient both set @opts
either to the options Hash passed to the constructor or and empty Hash
if no options Hash was passed.  Having both set @opts is problematic
because RoachClient is a subclass of Client, so after RoachClient
removes the options Hash from the argument list, stores it in @opts, and
then calls Client's constructor via super, the Client constructor then
sets @opts to an empty Hash, thereby obliterating the desired Hash!

This problem is fixed by setting @opts only if it has not yet been set.