Allow indexing TenGE ARP table by IP address
The TenGE ARP table is only 256 entries long.  The index for an entry
is the last octect of the corresponding IP address.  The index passed in
is now ANDed with 0xff to ensure that only the last 8 bits are used.
This allows one to access the ARP table by a 32-bit IP address
instead of having to mask off the upper bytes.

Instead of "tenge[ip&0xff]", users can now use "tenge[ip]" to get the
correspinding ARP entry.