Allow nil alias in typemap to control method name
Specifying a nil alias name and at least one non-nil alias name in a
device typemap entry now causes the synamically created method to be
named after the first non-nil alias rather than the device name.  This
can be used, for example, to specify a custom method name for a device
whose name is the same as an existing Ruby method name or keyword.